R1.45: Support for BK 2x Electric heating added
Both HLT E1 and BK E1 can be sourced from the same outlet. Same is valid for HLT E2 and BK E2. Routine pwm_task() makes sure that if both heaters on the same outlet are enabled, only the BK heater is allowed to draw current (the HLT heater is disabled in that time-slot).